separate <Resource/> schedule in own <Schedule/> component
